Move loggers to their own directory
